Added README and restructured project
A README was written for the repository and the project structure was changed to make more sense.
The equations have been moved out of the main Rmarkdown file are called as a function from a tex file.
Biobliography and appendices have been properly added now with the `after_body.Rmd` file.
The abstract text has also been put into a seperate file that is imported into the title.tex instead of being fully typed out there.